+++ src/main/resources/shop/src/app/admin-subcategory/admin-subcategory.component.ts
@@ -7,6 +7,8 @@ import {CommonAbstractComponent} from '../common/common-abstract-component';
 import {CommonLanguageModel} from '../common/common-language.model';
 import {CategoriesModel} from '../data/models/categories.model';
 import {AdditionalLinks} from '../data/additional-links';
+import {NoItem} from '../common/common-language.interface';
+import {ConfirmDialogComponent} from '../dialogs/confirm-dialog/confirm-dialog.component';
 
 @Component({
   selector: 'app-admin-subcategory',
@@ -15,7 +17,7 @@ import {AdditionalLinks} from '../data/additional-links';
 })
 export class AdminSubcategoryComponent extends CommonAbstractComponent implements OnInit, OnDestroy {
 
-  public get allSubcategories(): Subcategory[] {
+  public get subcategories(): Subcategory[] {
     return this._categoriesModel.subCategories;
   }
 
@@ -23,6 +25,8 @@ export class AdminSubcategoryComponent extends CommonAbstractComponent implement
     new AdditionalLinks(this.language.addSubcategory, '/admin-categories/add-subcategory'),
     new AdditionalLinks(this.language.categories, '/admin-categories'),
   ];
+
+  public noItemType: NoItem = NoItem.subcategories;
 
   constructor(private _categoriesModel: CategoriesModel,
               private _dialog: MatDialog,
@@ -42,12 +46,18 @@ export class AdminSubcategoryComponent extends CommonAbstractComponent implement
   }
 
   public delete(subcategory: Subcategory): void {
-    const category = this._categoriesModel.categories.find(category => {
-      return category.subcategories.find(categorySub => categorySub.id === subcategory.id)
+    const dialogRef = this._dialog.open(ConfirmDialogComponent);
+
+    dialogRef.afterClosed().subscribe(confirmed => {
+      if (confirmed) {
+        const category = this._categoriesModel.categories.find(category => {
+          return category.subcategories.find(categorySub => categorySub.id === subcategory.id);
+        });
+        if (category) {
+          this._categoriesModel.deleteSubcategory(subcategory, category);
+        }
+      }
     });
-    if (category) {
-      this._categoriesModel.deleteSubcategory(subcategory, category);
-    }
   }
 
   public edit(subcategory: Subcategory): void {
